+++ src/app/Products/product-list/product-list.component.ts
@@ -1,5 +1,9 @@
 import { Component, OnInit } from '@angular/core';
-import { ProductModel } from '../product-model';
+import { CartModule } from 'src/app/Cards/cart.module';
+import { CartService } from 'src/app/Cards/cart.service';
+import { Category } from 'src/app/category';
+import { CartModel } from 'src/app/models/cart.model';
+import { ProductModel } from '../../models/product.model';
 import { ProductsService } from '../products.service';
 
 @Component({
@@ -8,23 +12,45 @@ import { ProductsService } from '../products.service';
   styleUrls: ['./product-list.component.scss']
 })
 export class ProductListComponent implements OnInit {
-  products:ProductModel[] = []
-  name:string = '';
+  products: ProductModel[] = []
+  name: string = '';
 
-  constructor(private productService : ProductsService) { }
+  constructor(private productService: ProductsService) { }
 
   ngOnInit(): void {
     this.products = this.productService.getProducts();
   }
 
   filter(): void {
-    if(this.name == ''){
+    if (this.name == '') {
       this.products = this.productService.getProducts();
     }
-    else{
+    else {
       this.products = this.productService.filterByName(this.name);
     }
 
+  }
+
+  onAddToCart(product: ProductModel): void {
+    const cart: CartModel = {
+      name: product.name,
+      description: product.description,
+      price: product.price,
+      category: Category.Products,
+      isAvailable: product.isAvailable,
+      quantity: 1
+    }
+
+    this.productService.publishData(cart);
+  }
+
+  onLikeProduct(product: ProductModel): void {
+    this.productService.likeProduct(product);
+    console.log(this.productService.getLikedProducts());
+  }
+  onDislikeProduct(product: ProductModel): void {
+    this.productService.dislikeProduct(product);
+    console.log(this.productService.getLikedProducts());
   }
 
 }