+++ src/app/Products/product/product.component.ts
@@ -1,4 +1,5 @@
-import { Component } from '@angular/core';
+import { Component, EventEmitter, HostListener, Input, OnInit, Output } from '@angular/core';
+import { ProductModel } from '../../models/product.model';
 
 @Component({
   selector: 'app-product',
@@ -7,10 +8,32 @@ import { Component } from '@angular/core';
 })
 export class ProductComponentComponent {
 
-  constructor() { }
+
+  @Input()
+  product!: ProductModel;
+  @Output()
+  addToCart: EventEmitter<ProductModel> = new EventEmitter<ProductModel>();
+
+  @Output()
+  likeProduct: EventEmitter<ProductModel> = new EventEmitter<ProductModel>();
+  @Output()
+  dislikeProduct: EventEmitter<ProductModel> = new EventEmitter<ProductModel>();
+
+  activeHeart: boolean = false;
 
   onAddToCart(): void {
-      console.log("product was bought");
+    console.log("product was bought");
+    this.addToCart.emit(this.product);
+  }
+  onLikeProduct(): void {
+    if (this.activeHeart) {
+      this.dislikeProduct.emit(this.product);
+    } else {
+      this.likeProduct.emit(this.product);
+    }
+    this.activeHeart = !this.activeHeart;
+
+
   }
 
 }