+++ src/app/app.module.ts
@@ -3,25 +3,19 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { FirstComponentComponent } from './first-component/first-component.component';
-import { ProductComponentComponent } from './Products/product/product.component';
-import { ProductListComponent } from './Products/product-list/product-list.component';
-import { CartListComponent } from './Cards/cart-list/cart-list.component';
-import { FormsModule } from '@angular/forms';
+import { CartModule } from './Cards/cart.module';
+import { ProductsModule } from './Products/products.module';
 
 @NgModule({
   declarations: [
     AppComponent,
-    FirstComponentComponent,
-    ProductComponentComponent,
-    ProductListComponent,
-    CartListComponent
-    
+
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
-    FormsModule
+    CartModule,
+    ProductsModule
   ],
   providers: [],
   bootstrap: [AppComponent]