+++ src/app/models/cart.model.ts
@@ -1,0 +1,9 @@
+export interface CartModel {
+    name: string ;
+      description: string ;
+      price: number ;
+      category :string;
+      isAvailable: boolean;
+      quantity:number;
+    }
+