+++ src/app/models/product.model.ts
@@ -5,5 +5,3 @@ name: string ;
   category :string;
   isAvailable: boolean;
 }
-
-// правильное название этого файла product.model.ts