+++ CHANGELOG.md
@@ -9,3 +9,24 @@
 - CartListComponent.
 - CartService.
 - AddFilter.
+## [1.1.0] - 2021-09-21
+### Added
+
+- CartItemComponent.
+- CartModule.
+- CartModel.
+- OrderModule.
+- ProductModule.
+- SharedModule.
+- HighlightDirective.
+- AddLikedItem.
+### Changed
+
+- ProductListComponent (@Input(), @Output().
+- ProductComponent (@Input(), @Output()).
+- CartService (added count and summary).
+- CartListComponent (onPush).
+- AppComponent (set the title).
+- SharedModule.
+- HighlightDirective.
+- AddLikedItem.